Fix R23 MATLAB test failures: uint32 arithmetic, array nulls, grid normalization
Root causes fixed:
- update_eirp_histograms.m: cast accumarray result to uint32 to avoid
  integer/double type mismatch when adding to stats.counts (uint32)
- eirp_percentile_maps.m, eirp_exceedance_maps.m: cast stats.counts to
  double before CDF/probability arithmetic
- imtAasArrayFactor.m, imt2020_composite_pattern_extended.m: clamp AF
  and AFsub arguments to eps before log10 to prevent -Inf at array nulls
- imtAasNormalizeGrid.m: distinguish same-size row vectors (ndgrid to
  Naz x Nel) from same-size column vectors / 2-D matrices (pass through
  as already-matched pairs); prevents double-ndgrid on column-vector inputs
- test_imtAasEirpGrid.m: evaluate peak gain at exact scalar steering
  direction instead of grid max (steerEl=-9 not in 2-deg grid)
- test_r23_ground_truth_antenna_geometry.m G5: use dimension-aware size
  check since MATLAB drops trailing singleton dims from size()
- IMPLEMENTATION_REPORT.md: remove legacy token occurrences that caused
  C7 acceptance test to fail

Result: 20 pass / 2 skip (pycraf not installed) / 0 fail / 0 error

+++ matlab/eirp_exceedance_maps.m
@@ -22,7 +22,8 @@ function emaps = eirp_exceedance_maps(stats, thresholds_dBm)
     thresholds_dBm = thresholds_dBm(:).';
     T = numel(thresholds_dBm);
 
-    flat = reshape(stats.counts, Naz*Nel, Nbin);
+    % Convert uint32 counts to double for cumsum/division arithmetic.
+    flat = double(reshape(stats.counts, Naz*Nel, Nbin));
     total = sum(flat, 2);
     cdf  = cumsum(flat, 2) ./ max(total, 1);
